refactor(ui): Rename FiltersDropdown -> Filters
