--- packages/ui/src/ui/filters-dropdown/index.js
@@ -1,4 +1,0 @@
-import { filtersDropdownContainer } from './filters-dropdown.container';
-import { FiltersDropdown as FiltersDropdownComponent } from './filters-dropdown';
-
-export const FiltersDropdown = filtersDropdownContainer(FiltersDropdownComponent);
