+++ packages/ui/src/ui/filters/index.js
@@ -1,0 +1,4 @@
+import { filtersDropdownContainer } from './filters.container';
+import { Filters as FiltersComponent } from './filters';
+
+export const Filters = filtersDropdownContainer(FiltersComponent);
