+++ packages/ui/src/ui/index.js
@@ -3,7 +3,7 @@ export { Container } from './container';
 export { Dropdown } from './dropdown';
 export { EmptySet } from './empty-set';
 export { FileName } from './file-name';
-export { FiltersDropdown } from './filters-dropdown';
+export { Filters } from './filters';
 export { HorizontalBarChart } from './horizontal-bar-chart';
 export { Icon } from './icon';
 export { Loader } from './loader';
